+++ ios-device-lib/Classes/Terminals/UPA/HpsUpaDevice.m
@@ -48,6 +48,38 @@
     }];
 }
 
+-(void) upaPing:(void (^)(id<IHPSDeviceResponse> payload))responseBlock
+{
+    [self ping:^(id<IHPSDeviceResponse> payload, NSError *error) {
+        if (error) {
+            responseBlock(nil);
+        } else {
+            responseBlock(payload);
+        }
+    }];
+}
+
+-(void) upaAppRestart:(void (^)(id<IHPSDeviceResponse> payload))responseBlock
+{
+    [self restart:^(id<IHPSDeviceResponse>payload, NSError *error) {
+        if (error) {
+            responseBlock(nil);
+        } else {
+            responseBlock(payload);
+        }
+    }];
+}
+
+-(void) upaRebootDevice:(void (^)(id<IHPSDeviceResponse> payload))responseBlock
+{
+    [self reboot:^(id<IHPSDeviceResponse>payload, NSError *error) {
+        if (error) {
+            responseBlock(nil);
+        } else {
+            responseBlock(payload);
+        }
+    }];
+}
 /// method to force close communication early
 /// should only be used during administrative
 /// commands like pings (and not transaction requests)
@@ -121,12 +153,56 @@
     HpsUpaRequest* request = [[HpsUpaRequest alloc] init];
     request.message = @"MSG";
     request.data = [[HpsUpaCommandData alloc] init];
-    request.data.command = @"Ping";
+    request.data.command = UPA_MSG_ID_toString[ UPA_MSG_ID_PING ];
     request.data.EcrId = @"13";
     request.data.requestId = @"123";
 
     id<IHPSDeviceMessage> data = [HpsTerminalUtilities BuildRequest:request.JSONString withFormat:format];
+    NSLog(@"request json = \n  : %@", [request JSONString]);
+    [self.interface send:data andUPAResponseBlock:^(JsonDoc *data, NSError *error) {
+        if (error) {
+            dispatch_async(dispatch_get_main_queue(), ^{
+                responseBlock(nil, error);
+            });
+            return;
+        }
 
+        //done
+        NSString *dataview = [data toString];
+        NSLog(@"response json = \n  : %@", dataview);
+        HpsUpaResponse *response;
+
+        @try {
+            //parse data
+
+            response = [[HpsUpaResponse alloc]initWithJSONDoc:data];
+            dispatch_async(dispatch_get_main_queue(), ^{
+                responseBlock(response, nil);
+            });
+        } @catch (NSException *exception) {
+            NSDictionary *userInfo = @{NSLocalizedDescriptionKey: [exception description]};
+            NSError *error = [NSError errorWithDomain:self->errorDomain
+                                                 code:CocoaError
+                                             userInfo:userInfo];
+
+            dispatch_async(dispatch_get_main_queue(), ^{
+                responseBlock(nil, error);
+            });
+        }
+    }];
+}
+
+- (void) restart:(void(^)(id <IHPSDeviceResponse>, NSError*))responseBlock
+{
+    HpsUpaRequest* request = [[HpsUpaRequest alloc] init];
+    request.message = @"MSG";
+    request.data = [[HpsUpaCommandData alloc] init];
+    request.data.command = UPA_MSG_ID_toString[ UPA_MSG_ID_RESTART ];
+    request.data.EcrId = @"13";
+    request.data.requestId = @"123";
+
+    id<IHPSDeviceMessage> data = [HpsTerminalUtilities BuildRequest:request.JSONString withFormat:format];
+    NSLog(@"request json = \n  : %@", [request JSONString]);
     [self.interface send:data andUPAResponseBlock:^(JsonDoc *data, NSError *error) {
         if (error) {
             dispatch_async(dispatch_get_main_queue(), ^{
@@ -178,7 +254,7 @@
     request.data.requestId = @"123";
 
     id<IHPSDeviceMessage> data = [HpsTerminalUtilities BuildRequest:request.JSONString withFormat:format];
-
+    NSLog(@"request json = \n  : %@", [request JSONString]);
     [self.interface send:data andUPAResponseBlock:^(JsonDoc *data, NSError *error) {
         if (error) {
             dispatch_async(dispatch_get_main_queue(), ^{
@@ -221,9 +297,12 @@
     request.data.command = UPA_MSG_ID_toString[ UPA_MSG_ID_RESET ];
     request.data.EcrId = @"13";
     request.data.requestId = @"123";
+    request.data.data = [[HpsUpaData alloc] init];
+    request.data.data.params = [[HpsUpaParams alloc] init];
+    request.data.data.params.displayOption = @"1";
 
     id<IHPSDeviceMessage> data = [HpsTerminalUtilities BuildRequest:request.JSONString withFormat:format];
-
+    NSLog(@"request json = \n  : %@", [request JSONString]);
     [self.interface send:data andUPAResponseBlock:^(JsonDoc *data, NSError *error) {
         if (error) {
             dispatch_async(dispatch_get_main_queue(), ^{